analyst/requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==15.2.0
billiard==3.3.0.23
celery==3.1.20
cffi==1.5.2
cryptography==1.3
cssselect==0.9.1
decorator==4.0.9
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.8.8
django-bootstrap-pagination==1.6.1
django-celery==3.1.17
django-contrib-comments==1.6.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-dynamic-scraper==0.10.0
django-haystack==2.4.1
django-markdown-deux==1.0.5
django-markdownx==1.4.2
django-registration==2.0.4
django-taggit==0.18.0
enum34==1.1.2
future==0.15.2
gunicorn==19.4.5
idna==2.0
ipaddress==1.0.16
jsonpath-rw==1.4.0
kombu==3.0.33
lxml==3.6.0
Markdown==2.6.6
markdown2==2.3.1
Pillow==3.1.0
ply==3.8
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.0.0
pysolr==3.4.0
pytz==2016.2
queuelib==1.4.2
requests==2.9.1
Scrapy==1.0.5
scrapy-djangoitem==1.0.0
service-identity==16.0.0
six==1.10.0
sqlparse==0.1.19
static3==0.6.1
Twisted==16.0.0
w3lib==1.13.0
wheel==0.26.0
whitenoise==2.0.6
Whoosh==2.7.3
zope.interface==4.1.3
